+++ plugins/com.google.cloud.tools.eclipse.sdk.test/src/com/google/cloud/tools/eclipse/sdk/internal/CloudSdkInstallJobTest.java
@@ -16,18 +16,94 @@
 
 package com.google.cloud.tools.eclipse.sdk.internal;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertTrue;
 
+import java.util.concurrent.Semaphore;
+import org.eclipse.core.runtime.IProgressMonitor;
+import org.eclipse.core.runtime.IStatus;
+import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.jobs.Job;
+import org.eclipse.ui.console.MessageConsoleStream;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
 
+@RunWith(MockitoJUnitRunner.class)
 public class CloudSdkInstallJobTest {
 
-  @Test
-  public void testRun() throws InterruptedException {
-    Job job = new CloudSdkInstallJob(null);
-    job.schedule();
-    job.join();
-    assertTrue(job.getResult().isOK());
+  @Mock private MessageConsoleStream consoleStream;
+
+  private CloudSdkInstallJob installJob;
+
+  @Before
+  public void setUp() {
+    installJob = new FakeInstallJob(consoleStream, false /* blockOnStart */);
   }
+
+  @After
+  public void tearDown() {
+    assertEquals(Job.NONE, installJob.getState());
+  }
+
+  @Test
+  public void testBelongsTo() {
+    assertTrue(installJob.belongsTo(CloudSdkInstallJob.CLOUD_SDK_MODIFY_JOB_FAMILY));
+  }
+
+  @Test
+  public void testMutexRuleSet() {
+    assertEquals(CloudSdkInstallJob.MUTEX_RULE, installJob.getRule());
+  }
+
+  @Test
+  public void testRun_mutualExclusion() throws InterruptedException {
+    FakeInstallJob job1 = new FakeInstallJob(null, true /* blockOnStart */);
+    FakeInstallJob job2 = new FakeInstallJob(null, true);
+
+    job1.schedule();
+    while (job1.getState() != Job.RUNNING) {
+      Thread.sleep(10);
+    }
+
+    job2.schedule();
+    // Incomplete test, but if it ever fails, something is surely broken.
+    assertNotEquals(Job.RUNNING, job2.getState());
+
+    job1.unblock();
+    job2.unblock();
+    job1.join();
+    job2.join();
+  }
+
+  private class FakeInstallJob extends CloudSdkInstallJob {
+
+    private final Semaphore blocker = new Semaphore(0);
+    private final boolean blockOnStart;
+
+    private FakeInstallJob(MessageConsoleStream consoleStream, boolean blockOnStart) {
+      super(consoleStream);
+      this.blockOnStart = blockOnStart;
+    }
+
+    @Override
+    protected IStatus run(IProgressMonitor monitor) {
+      try {
+        if (blockOnStart) {
+          blocker.acquire();
+        }
+        return Status.OK_STATUS;
+      } catch (InterruptedException e) {
+        return Status.CANCEL_STATUS;
+      }
+    }
+
+    private void unblock() {
+      blocker.release();
+    }
+  };
 }
