+++ plugins/com.google.cloud.tools.eclipse.sdk/src/com/google/cloud/tools/eclipse/sdk/internal/CloudSdkInstallJob.java
@@ -17,20 +17,39 @@
 package com.google.cloud.tools.eclipse.sdk.internal;
 
 import com.google.cloud.tools.eclipse.sdk.Messages;
+import com.google.cloud.tools.eclipse.util.jobs.MutexRule;
+import com.google.common.annotations.VisibleForTesting;
+import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Status;
+import org.eclipse.core.runtime.jobs.Job;
 import org.eclipse.ui.console.MessageConsoleStream;
 
-public class CloudSdkInstallJob extends CloudSdkModifyJob {
+public class CloudSdkInstallJob extends Job {
+
+  public static final Object CLOUD_SDK_MODIFY_JOB_FAMILY = new Object();
+
+  /** Scheduling rule to prevent running {@code CloudSdkModifyJob} concurrently. */
+  @VisibleForTesting
+  static final MutexRule MUTEX_RULE = new MutexRule();
+
+  private final MessageConsoleStream consoleStream;
 
   public CloudSdkInstallJob(MessageConsoleStream consoleStream) {
-    super(Messages.getString("installJobName"), consoleStream); //$NON-NLS-1$
+    super(Messages.getString("installJobName"));
+    this.consoleStream = consoleStream;
+    setRule(MUTEX_RULE);
   }
 
   @Override
-  protected IStatus modifySdk() {
-    // TODO(chanseok): to be implemented: https://github.com/GoogleCloudPlatform/google-cloud-eclipse/issues/2753
-    return Status.OK_STATUS;
+  public boolean belongsTo(Object family) {
+    return super.belongsTo(family) || family == CLOUD_SDK_MODIFY_JOB_FAMILY;
   }
 
+  @Override
+  protected IStatus run(IProgressMonitor monitor) {
+    // TODO(chanseok): to be implemented: https://github.com/GoogleCloudPlatform/google-cloud-eclipse/issues/2753
+    consoleStream.println("to be implemented");
+    return Status.OK_STATUS;
+  }
 }
