+++ plugins/com.google.cloud.tools.eclipse.sdk/src/com/google/cloud/tools/eclipse/sdk/messages.properties
@@ -1,3 +1,1 @@
 installJobName=Installing the Cloud SDK...
-startModifying=Installing/upgrading the Cloud SDK... (may take several minutes)
-sdkModificationLocked=Cannot install/upgrade because the Cloud SDK is in use.
